[FFHRS-2] Rename coventry logo svg
